Fix unwanted character sliding on slopes and add new ramp obstical
The character used to slide down slopes with a low enough incline to be
the floor.

The character used to run up (or slide up if in the Air state) slopes
that are too steep to be the floor.

New bad behaviour:
- On slopes with an incline equal to max_floor_angle, the character
makes a slight hop after running up and looses their grounding when
running down them.
- When the character runs into a steep slope they perform short hops due
to sliding upwards. Ideally the player would stay on the ground (no
sliding up the slope). This can be exploited to jump higher.

A new ramp obstical (Rampy) was added to the main Game scene to test the
slope behaviour.

Closes: #25